requirements/base-linux.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile base-linux.in
#
aiohttp==3.8.1
    # via gradio
aiosignal==1.2.0
    # via aiohttp
analytics-python==1.4.0
    # via gradio
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
backoff==1.10.0
    # via analytics-python
bcrypt==4.0.0
    # via paramiko
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via uvicorn
concrete-compiler==0.6.0
    # via concrete-numpy
concrete-numpy==0.5.0
    # via -r base-linux.in
cryptography==37.0.4
    # via paramiko
cycler==0.11.0
    # via matplotlib
fastapi==0.80.0
    # via gradio
ffmpy==0.3.0
    # via gradio
fire==0.4.0
    # via medmnist
fonttools==4.37.0
    # via matplotlib
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2022.7.1
    # via gradio
gradio==3.1.7
    # via -r base-linux.in
h11==0.12.0
    # via
    #   gradio
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via gradio
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
imageio==2.21.1
    # via scikit-image
jinja2==3.1.2
    # via gradio
joblib==1.1.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
linkify-it-py==1.0.3
    # via markdown-it-py
loguru==0.5.3
    # via concrete-numpy
markdown-it-py[linkify,plugins]==2.1.0
    # via
    #   gradio
    #   mdit-py-plugins
markupsafe==2.1.1
    # via jinja2
matplotlib==3.5.3
    # via
    #   concrete-numpy
    #   gradio
mdit-py-plugins==0.3.0
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
medmnist==2.1.0
    # via -r base-linux.in
monotonic==1.6
    # via analytics-python
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
networkx==2.8.6
    # via
    #   concrete-numpy
    #   scikit-image
numpy==1.23.2
    # via
    #   -r base-linux.in
    #   concrete-compiler
    #   concrete-numpy
    #   gradio
    #   imageio
    #   matplotlib
    #   medmnist
    #   onnx
    #   pandas
    #   pywavelets
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
    #   torchvision
onnx==1.12.0
    # via -r base-linux.in
orjson==3.7.12
    # via gradio
packaging==21.3
    # via
    #   matplotlib
    #   scikit-image
pandas==1.4.3
    # via
    #   gradio
    #   medmnist
paramiko==2.11.0
    # via gradio
pillow==9.2.0
    # via
    #   concrete-numpy
    #   gradio
    #   imageio
    #   matplotlib
    #   medmnist
    #   scikit-image
    #   torchvision
protobuf==3.20.1
    # via onnx
pycparser==2.21
    # via cffi
pycryptodome==3.15.0
    # via gradio
pydantic==1.9.2
    # via
    #   fastapi
    #   gradio
pydub==0.25.1
    # via gradio
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   analytics-python
    #   matplotlib
    #   pandas
python-multipart==0.0.5
    # via gradio
pytz==2022.2.1
    # via pandas
pywavelets==1.3.0
    # via scikit-image
pyyaml==6.0
    # via concrete-compiler
requests==2.28.1
    # via
    #   analytics-python
    #   gradio
    #   torchvision
rfc3986[idna2008]==1.5.0
    # via httpx
scikit-image==0.19.3
    # via medmnist
scikit-learn==1.1.2
    # via medmnist
scipy==1.9.0
    # via
    #   -r base-linux.in
    #   scikit-image
    #   scikit-learn
six==1.16.0
    # via
    #   analytics-python
    #   fire
    #   paramiko
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.19.1
    # via fastapi
termcolor==1.1.0
    # via fire
threadpoolctl==3.1.0
    # via scikit-learn
tifffile==2022.8.12
    # via scikit-image
torch==1.12.1
    # via
    #   -r base-linux.in
    #   concrete-numpy
    #   medmnist
    #   torchvision
torchvision==0.13.1
    # via
    #   -r base-linux.in
    #   medmnist
tqdm==4.64.0
    # via medmnist
typing-extensions==4.3.0
    # via
    #   onnx
    #   pydantic
    #   starlette
    #   torch
    #   torchvision
uc-micro-py==1.0.1
    # via linkify-it-py
urllib3==1.26.12
    # via requests
uvicorn==0.18.3
    # via gradio
websockets==10.3
    # via gradio
yarl==1.8.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
